django/requirements.txt
Django==5.0.1
reportlab==4.0.9
Pillow==10.1.0
python-dotenv==1.0.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
